Fix api and auth and implement the vuex store
