+++ resources/assets/js/api/axios_interceptors.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import router from '@/router';
-import api from './api';
+import api from './';
 
 axios.interceptors.request.use(request => {
     const token = api.auth.token();
