+++ resources/assets/js/store/index.js
@@ -5,17 +5,17 @@ Vue.use(Vuex);
 
 import auth from './modules/auth';
 
-// import createPersistedState from 'vuex-persistedstate';
+import createPersistedState from 'vuex-persistedstate';
 
 const modules = {
     auth
 };
 
-// const plugins = [
-//     createPersistedState()
-// ];
+ const plugins = [
+     createPersistedState()
+ ];
 
 export default new Vuex.Store({
     modules,
-    // plugins
+    plugins
 });
