+++ resources/assets/js/stories/auth/Login.vue
@@ -20,7 +20,7 @@
 </template>
 
 <script>
-    import api from '@/api/api';
+    import api from '@/api';
 
     export default {
         name: "Login",
